feat: started workin on the tests

+++ server/src/test/mountain.test.js
@@ -1,31 +1,11 @@
-import { test, describe, after, beforeEach } from 'node:test'
-import mongoose from 'mongoose'
+import { test, describe, after, beforeEach, before } from 'node:test'
 import supertest from 'supertest'
 import app from '../../app.js'
 import assert from 'node:assert'
 
 import Location from '../models/location.js'
-import Hill from '../models/hill.js'
 import Mountain from '../models/mountain.js'
-
-import RFIDTag from '../models/rfidTag.js'
-import Event from '../models/event.js'
-import Team from '../models/team.js'
-import Participant from '../models/participant.js'
-import Lap from '../models/lap.js'
-
-export const emptyTestDB = async () => {
-  await Promise.all([
-    Location.deleteMany({}),
-    Hill.deleteMany({}), 
-    Mountain.deleteMany({}),
-    RFIDTag.deleteMany({}),
-    Event.deleteMany({}),
-    Team.deleteMany({}),
-    Participant.deleteMany({}),
-    Lap.deleteMany({}),
-  ])
-}
+import { authToken, closeDBConnection, emptyTestDB } from './testHelper.js'
 
 const api = supertest(app)
 
@@ -43,7 +23,6 @@ beforeEach(async () => {
     name: 'Test Park', address: '1 Test St', city: 'Testville', provState: 'TS', country: 'Testland'
   }).save()
   aLocationId = location._id
-  console.log('🚀 ~ beforeEach ~ aLocationId:', aLocationId)
   await Mountain.insertMany(initialMountainsData)
 })
 
@@ -51,12 +30,13 @@ describe('Mountains API (/api/mountains)', () => {
   test('mountains are returned as json', async () => {
     await api
       .get('/api/mountains')
+      .set('Authorization', `bearer ${authToken}`)
       .expect(200)
       .expect('Content-Type', /application\/json/)
   })
 
   test('all mountains are returned', async () => {
-    const response = await api.get('/api/mountains')
+    const response = await api.get('/api/mountains').set('Authorization', `bearer ${authToken}`)
     assert.strictEqual(response.body.length, initialMountainsData.length)
   })
 
@@ -68,12 +48,12 @@ describe('Mountains API (/api/mountains)', () => {
     }
 
     await api
-      .post('/api/mountains')
+      .post('/api/mountains').set('Authorization', `bearer ${authToken}`)
       .send(newMountain)
       .expect(201)
       .expect('Content-Type', /application\/json/)
 
-    const response = await api.get('/api/mountains')
+    const response = await api.get('/api/mountains').set('Authorization', `bearer ${authToken}`)
     const mountainNames = response.body.map(m => m.name)
 
     assert.strictEqual(response.body.length, initialMountainsData.length + 1)
@@ -81,28 +61,28 @@ describe('Mountains API (/api/mountains)', () => {
   })
 
   test('a mountain can be updated', async () => {
-    const mountains = await api.get('/api/mountains')
+    const mountains = await api.get('/api/mountains').set('Authorization', `bearer ${authToken}`)
     const mountainToUpdate = mountains.body[0]
     const payload = { ...mountainToUpdate, name: 'UPDATED Everest' }
 
-    await api.put(`/api/mountains/${mountainToUpdate.id}`).send(payload).expect(200)
+    await api.put(`/api/mountains/${mountainToUpdate.id}`).set('Authorization', `bearer ${authToken}`).send(payload).expect(200)
 
-    const res = await api.get(`/api/mountains/${mountainToUpdate.id}`)
+    const res = await api.get(`/api/mountains/${mountainToUpdate.id}`).set('Authorization', `bearer ${authToken}`)
     assert.strictEqual(res.body.name, 'UPDATED Everest')
   })
 
   test('a mountain can be deleted', async () => {
-    const mountains = await api.get('/api/mountains')
+    const mountains = await api.get('/api/mountains').set('Authorization', `bearer ${authToken}`)
     const mountainToDelete = mountains.body[0]
 
-    await api.delete(`/api/mountains/${mountainToDelete.id}`).expect(204)
+    await api.delete(`/api/mountains/${mountainToDelete.id}`).set('Authorization', `bearer ${authToken}`).expect(204)
     
-    const finalMountains = await api.get('/api/mountains')
+    const finalMountains = await api.get('/api/mountains').set('Authorization', `bearer ${authToken}`)
     assert.strictEqual(finalMountains.body.length, initialMountainsData.length - 1)
   })
 })
 
 
 after(async () => {
-  await mongoose.connection.close()
+  await closeDBConnection()
 })
